+++ jquery.reel.js
@@ -196,6 +196,11 @@
         //
         wheelable:           true,
 
+        // #### `orientable` Option ####
+        // `Boolean`, IN DEVELOPMENT
+        //
+        orientable:         false,
+
 
         // ### Order of Images ######
         //
@@ -1011,6 +1016,7 @@
                     no_bias();
                     pool.unbind(on.pool);
                     pools.unbind(pns);
+                    $(window).unbind(ns);
                   },
 
                   // ### `setup` Event ######
@@ -1184,8 +1190,10 @@
                       opening= set(_opening_, false),
                       evnt= _tick_+dot(_opening_)
                     pool.unbind(evnt, on.pool[evnt]);
+                    touchy && opt.orientable && $(window).bind(_deviceorientation_, orient);
                     if (opt.delay > 0) delay= setTimeout(function(){ t.trigger('play') }, opt.delay * 1000)
                     else t.trigger('play');
+                    function orient(e){ return t.trigger('orient', [gyro(e).alpha, gyro(e).beta, gyro(e).gamma, e]) && e.give }
                   },
 
                   // -----------------------
@@ -1390,6 +1398,24 @@
                     ev && (ev.give = false);
                     unidle();
                     t.trigger('up', [ev]);
+                  },
+
+                  // ### `orient` Event ######
+                  // `Event`, IN DEVELOPMENT
+                  //
+                  // Maps Reel to device orientation event which is provided by some touch enabled devices
+                  // with gyroscope inside. Event handler receives all three device orientation angles 
+                  // in arguments. This event:
+                  //
+                  // - maps alpha angle directly to `fraction`
+                  //
+                  orient: function(e, alpha, beta, gamma, ev){
+                    if (!slidable || operated) return;
+                    oriented= true;
+                    var
+                      alpha_fraction= alpha / 360
+                      fraction= set(_fraction_, +((opt.stitched || opt.cw ? 1 - alpha_fraction : alpha_fraction)).toFixed(2))
+                    slidable = false;
                   },
 
                   // ------------------
@@ -1801,7 +1827,7 @@
                     var
                       direction= get(_cwish_) * negative_when(1, get(_backwards_)),
                       ticks= get(_ticks_),
-                      step= (!get(_playing_) || !ticks ? velocity : abs(get(_speed_)) + velocity) / leader(_tempo_),
+                      step= (!get(_playing_) || oriented || !ticks ? velocity : abs(get(_speed_)) + velocity) / leader(_tempo_),
                       fraction= set(_fraction_, get(_fraction_) - step * direction),
                       ticks= !opt.duration ? ticks : ticks > 0 && set(_ticks_, ticks - 1)
                     !ticks && get(_playing_) && t.trigger('stop');
@@ -1844,6 +1870,7 @@
               },
               panned= 0,
               wheeled= false,
+              oriented= false,
               delay, // openingDone's delayed play pointer
 
               // - Constructors of UI elements
@@ -2427,6 +2454,7 @@
     _mouseleave_= _mouse_+'leave'+pns, _mousemove_= _mouse_+'move'+pns, _mouseup_= _mouse_+'up'+pns,
     _mousewheel_= _mouse_+'wheel'+ns, _tick_= 'tick'+ns, _touchcancel_= _touch_+'cancel'+pns,
     _touchend_= _touch_+'end'+pns, _touchstart_= _touch_+'start'+ns, _touchmove_= _touch_+'move'+pns,
+    _deviceorientation_= 'deviceorientation'+ns,
 
     // And some other frequently used Strings.
     //
@@ -2497,6 +2525,7 @@
   function min_max(minimum, maximum, number){ return max(minimum, min(maximum, number)) }
   function negative_when(value, condition){ return abs(value) * (condition ? -1 : 1) }
   function finger(e){ return touchy ? e.touch || e.originalEvent.touches[0] : e }
+  function gyro(e){ return e.originalEvent }
   function px(value){ return value === undefined ? 0 : typeof value == _string_ ? value : value + 'px' }
   function hash(value){ return '#' + value }
   function pad(string, len, fill){ while (string.length < len) string= fill + string; return string }